requirements.txt
torch>=1.7.0
transformers>=4.5.0
pandas>=1.0.0
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
huggingface_hub[hf_xet]>=0.23.0